2018_07_04
클라이언트 PlayerInput 수정(키 이벤트 수정)
서버 타이머 스레드 작동점검(초당 패킷 송신 확인 수 확인)
캐릭터 모션 수정(점프)

+++ 최규화/Mod_Server/Server/PhysicsEngine/MyMiniPysicsEngine.cpp
@@ -174,7 +174,7 @@ PhysicsPoint::~PhysicsPoint()
 {
 }
 
-void PhysicsPoint::integrate(__int64 DeltaTime, XMFLOAT4* ObjPos, XMFLOAT3* ObjVel)
+void PhysicsPoint::integrate(double DeltaTime, XMFLOAT4* ObjPos, XMFLOAT3* ObjVel)
 {
 
 	if (InverseMass <= 0.0f)
